Started rewriting to store variable arcs in OCLPMs.
Now storing variable arcs inside of OCLPMs as a map from the place ids of all isomorphic places to the activities with which they have variable arcs. To make checking if a place is inside a model and more, the isomorphism check swaps identical (isomorphic and same type) with the place from the place set. This ensures that the places inside the model actually use the ids of the places from the place set as they could have been changed during the LPM discovery.